Close #12 - Fix bdc-db migration column types

+++ bdc_db/models/raster_size_schema.py
@@ -6,9 +6,9 @@ class RasterSizeSchema(BaseModel):
     __tablename__ = 'raster_size_schemas'
 
     id = Column(String(60), primary_key=True)
-    raster_size_x = Column(Float(50), nullable=False)
-    raster_size_y = Column(Float(50), nullable=False)
-    raster_size_t = Column(Float(50))
-    chunk_size_x = Column(Float(50), nullable=False)
-    chunk_size_y = Column(Float(50), nullable=False)
-    chunk_size_t = Column(Float(50))
+    raster_size_x = Column(Float(53), nullable=False)
+    raster_size_y = Column(Float(53), nullable=False)
+    raster_size_t = Column(Float(53))
+    chunk_size_x = Column(Float(53), nullable=False)
+    chunk_size_y = Column(Float(53), nullable=False)
+    chunk_size_t = Column(Float(53))
